getopt replaced by ezOptionParser-0.2.2
Command line option handling now used ezOptionParser-0.2.2. The
benefit is that it alows creation and use of configure files to avoid
lengthy command lines. Options on command line take precedence, then
options specified in --read-config filename, then options specified in
the file "config" if that exists in the launch folder.  Disadvantage is that
it does not allow options of the form --cell-permute=2 but requres space
separation between option name and argument.

Arguments for multisend, bin queue, and spike compression.

--ms-phases ARG          Number of multisend phases, 1 or 2. (2)
--ms-subintervals ARG    Number of multisend subintervals, 1 or 2. (2)
--multisend              Use Multisend spike exchange instead of Allgather.
--binqueue               Use bin queue.
--spkcompress ARG        Spike compression. Up to ARG are exchanged during MPI_Allgather. (0)